nedii/api/requirements.txt
asgiref==3.4.1
bcrypt==3.2.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.5
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.12
django-3-jet==1.0.8
django-colorfield==0.6.3
django-cors-headers==3.11.0
django-environ==0.8.1
django-filter==21.1
django-resized==0.3.11
django-tinymce==3.4.0
django-tinymce4-lite==1.8.0
djangorestframework==3.13.1
djangorestframework-jsonapi==5.0.0
djangorestframework-simplejwt==5.1.0
drf-yasg==1.20.0
environ==1.0
gunicorn==20.1.0
idna==3.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
jsmin==2.0
MarkupSafe==2.0.1
packaging==21.0
Pillow==8.3.2
psycopg2==2.9.3
pycparser==2.20
pyenchant==3.2.2
PyJWT==2.1.0
pyparsing==2.4.7
pytz==2021.1
requests==2.26.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
uritemplate==3.0.1
urllib3==1.26.6
